+++ noodles-bam/CHANGELOG.md
@@ -27,6 +27,12 @@
     If the start position or CIGAR is changed, the bin number in the record may
     be incorrect.
 
+### Removed
+
+  * bam/record: Remove bin number (`Record::bin`).
+
+    This can be calculated from the alignment start and end instead.
+
 ## 0.16.0 - 2022-03-02
 
 ### Added